Create example  dart
In the main menu, we wrote code that contains a list called AA, which contains numbers. Then we write a print command that is as long as the list, so the number five is printed first. Then after that, there is a command to print the entire set of the list using sort, then the print command that prints All items in the list.